Move Software class from DTO to Models
