+++ src/DevSpector.SDK/Interfaces/IDevicesEditor.cs
@@ -1,5 +1,6 @@
 using System.Threading.Tasks;
 using DevSpector.SDK.DTO;
+using DevSpector.SDK.Models;
 
 namespace DevSpector.SDK
 {
